moved WaitForSingleObject into kernel32.synchronization
