+++ pywincffi/kernel32/handle.py
@@ -3,7 +3,7 @@ Handles
 -------
 
 A module containing general functions for working with handle
-objects.
+objects.  The functions provided here are part of the ``kernel32`` library.
 """
 
 from six import integer_types
@@ -87,39 +87,6 @@ def CloseHandle(hObject):
 
     code = library.CloseHandle(hObject)
     error_check("CloseHandle", code=code, expected=Enums.NON_ZERO)
-
-
-def WaitForSingleObject(hHandle, dwMilliseconds):
-    """
-    Waits for the specified object to be in a signaled state
-    or for ``dwMiliseconds`` to elapse.
-
-    .. seealso::
-
-        https://msdn.microsoft.com/en-us/library/ms687032
-
-    :param handle hHandle:
-        The handle to wait on.
-
-    :param int dwMilliseconds:
-        The time-out interval.
-    """
-    input_check("hHandle", hHandle, Enums.HANDLE)
-    input_check("dwMilliseconds", dwMilliseconds, integer_types)
-
-    ffi, library = dist.load()
-    result = library.WaitForSingleObject(
-        hHandle, ffi.cast("DWORD", dwMilliseconds)
-    )
-
-    if result == library.WAIT_FAILED:
-        raise WindowsAPIError(
-            "WaitForSingleObject", "Wait Failed", ffi.getwinerror()[-1],
-            return_code=result, expected_return_code="not %s" % result)
-
-    error_check("WaitForSingleObject")
-
-    return result
 
 
 def GetHandleInformation(hObject):
